Removed cdn link for fontAwesome in favor of downloaded files. The webfonts folder is needed for this purpose.
Slightly refactored both filterResources and AddSearchResources to remove lots of
redundant html. (inside the if statements)

When a user clicks on a card for a resource the resource id will be captured and
calendar events will be filtered by that resource. the resource card will also 
highlight to show it has been clicked. 

When adding an event in the week view the user can choose to make the event recurring
by up to 4 weeks. These are successfully added to the data base. 
(Nice to have would be deleting recurring event would give the option to delete all others,
currently not possible. Not sure if it would be possible without changing db. )

Current Known Issue: Back button in resources screen causes error.